Rename index.tsx to VideoItem.tsx
